api/requirements.txt
connexion[swagger-ui]
pytest
boto3
moto
flask_api
PyJWT
cryptography
serverless_wsgi
Flask-Cors
strict-rfc3339
pytz
PyYAML
